refactor(middleware): use public bucket handler directly
