+++ package.json
@@ -42,7 +42,7 @@
     "@ucanto/client": "^9.0.1",
     "@ucanto/principal": "^9.0.1",
     "@ucanto/transport": "^9.1.1",
-    "@web3-storage/blob-fetcher": "^2.4.4-rc.0",
+    "@web3-storage/blob-fetcher": "^2.4.4-rc.1",
     "@web3-storage/capabilities": "^17.4.1",
     "@web3-storage/gateway-lib": "^5.1.2",
     "dagula": "^8.0.0",
